Application du squelette au suivi instruction
